Fixed big bug in unrealFloatFromPng

+++ neural_wrappers/readers/datasets/carla/utils.py
@@ -3,8 +3,9 @@ import transforms3d.euler as txe
 from ....utilities import npCloseEnough, npGetInfo
 
 def unrealFloatFromPng(x : np.ndarray) -> np.ndarray:
+	x = x.astype(np.float32)
 	x = (x[..., 0] + x[..., 1] * 256 + x[..., 2] * 256 * 256) / (256 * 256 * 256 - 1)
-	x = np.float32(x)
+	x = x.astype(np.float32)
 	return x
 
 def unrealPngFromFloat(x : np.ndarray) -> np.ndarray:
